pay method done for BankClient class

+++ src/main/java/bank/BankAccount.java
@@ -38,7 +38,7 @@ public class BankAccount {
         return this.password;
     }
 
-    public synchronized long getMoney() {
+    public synchronized int getMoney() {
         return this.deposit;
     }
